+++ packages/bluesky/test/src/feeds/feeds_service_test.dart
@@ -6,7 +6,20 @@
 import 'package:atproto/atproto.dart';
 import 'package:atproto_core/atproto_core.dart';
 import 'package:atproto_test/atproto_test.dart' as atp_test;
-import 'package:bluesky/bluesky.dart';
+import 'package:bluesky/src/entities/actor_feeds.dart';
+import 'package:bluesky/src/entities/feed.dart';
+import 'package:bluesky/src/entities/feed_generator.dart';
+import 'package:bluesky/src/entities/feed_generators.dart';
+import 'package:bluesky/src/entities/likes.dart';
+import 'package:bluesky/src/entities/post_thread.dart';
+import 'package:bluesky/src/entities/posts.dart';
+import 'package:bluesky/src/entities/reposted_by.dart';
+import 'package:bluesky/src/entities/skeleton_feed.dart';
+import 'package:bluesky/src/feeds/feeds_service.dart';
+import 'package:bluesky/src/params/generator_param.dart';
+import 'package:bluesky/src/params/post_param.dart';
+import 'package:bluesky/src/params/strong_ref_param.dart';
+import 'package:bluesky/src/params/thread_param.dart';
 // 📦 Package imports:
 import 'package:test/test.dart';
 
@@ -783,7 +796,7 @@ void main() {
     });
   });
 
-  group('.findCustomFeed', () {
+  group('.findSkeletonFeed', () {
     test('normal case', () async {
       final feeds = FeedsService(
         atproto: ATProto(did: 'test', accessJwt: 'test'),
@@ -851,6 +864,81 @@ void main() {
 
       atp_test.expectRateLimitExceededException(
         () async => await feeds.findCustomFeed(
+            generatorUri: AtUri.parse(
+          'at://did:plc:tenurhgjptubkk5zf5qhi3og/app.bsky.feed.generator/h-privacy',
+        )),
+      );
+    });
+  });
+
+  group('.findCustomFeed', () {
+    test('normal case', () async {
+      final feeds = FeedsService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/feeds/data/find_skeleton_feed.json',
+        ),
+      );
+
+      final response = await feeds.findSkeletonFeed(
+        generatorUri: AtUri.parse(
+          'at://did:plc:tenurhgjptubkk5zf5qhi3og/app.bsky.feed.generator/h-privacy',
+        ),
+        limit: 10,
+        cursor: '1234',
+      );
+
+      expect(response, isA<XRPCResponse>());
+      expect(response.data, isA<SkeletonFeed>());
+    });
+
+    test('when unauthorized', () async {
+      final feeds = FeedsService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/data/error.json',
+          statusCode: 401,
+        ),
+      );
+
+      atp_test.expectUnauthorizedException(
+        () async => await feeds.findSkeletonFeed(
+          generatorUri: AtUri.parse(
+            'at://did:plc:tenurhgjptubkk5zf5qhi3og/app.bsky.feed.generator/h-privacy',
+          ),
+        ),
+      );
+    });
+
+    test('when rate limit exceeded', () async {
+      final feeds = FeedsService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/data/error.json',
+          statusCode: 429,
+        ),
+      );
+
+      atp_test.expectRateLimitExceededException(
+        () async => await feeds.findSkeletonFeed(
             generatorUri: AtUri.parse(
           'at://did:plc:tenurhgjptubkk5zf5qhi3og/app.bsky.feed.generator/h-privacy',
         )),